restartService --> forceStopService
JobService shouldn't need to be restarted, but simply killed.
Restarting the service presents requires an app to call the service
start intent, which right now was interfacer itself, and it got
caught in our Signature check (as it isn't Substratum's check).

When a new job intent is broadcasted, JobService should be launched
(restarted) automatically anyway.

Change-Id: If0bb66fa7481b1865686e0b1393a270a8c3df0ba